+++ tests/JsonApi/Hydrator/AbstractHydratorTest.php
@@ -2,7 +2,7 @@
 namespace WoohooLabsTest\Yin\JsonApi\Hydrator;
 
 use PHPUnit_Framework_TestCase;
-use WoohooLabs\Yin\JsonApi\Exception\ExceptionFactory;
+use WoohooLabs\Yin\JsonApi\Exception\DefaultExceptionFactory;
 use WoohooLabs\Yin\JsonApi\Hydrator\Relationship\ToManyRelationship;
 use WoohooLabs\Yin\JsonApi\Hydrator\Relationship\ToOneRelationship;
 use WoohooLabs\Yin\JsonApi\Request\Request;
@@ -23,7 +23,7 @@ class AbstractHydratorTest extends PHPUnit_Framework_TestCase
         ];
 
         $hydrator = $this->createHydrator();
-        $hydrator->hydrateForCreate($this->createRequest($body), new ExceptionFactory(), []);
+        $hydrator->hydrateForCreate($this->createRequest($body), new DefaultExceptionFactory(), []);
     }
 
     /**
@@ -39,7 +39,7 @@ class AbstractHydratorTest extends PHPUnit_Framework_TestCase
         ];
 
         $hydrator = $this->createHydrator("fox");
-        $hydrator->hydrateForCreate($this->createRequest($body), new ExceptionFactory(), []);
+        $hydrator->hydrateForCreate($this->createRequest($body), new DefaultExceptionFactory(), []);
     }
 
     /**
@@ -55,7 +55,7 @@ class AbstractHydratorTest extends PHPUnit_Framework_TestCase
         ];
 
         $hydrator = $this->createHydrator(["fox", "wolf"]);
-        $hydrator->hydrateForUpdate($this->createRequest($body), new ExceptionFactory(), []);
+        $hydrator->hydrateForUpdate($this->createRequest($body), new DefaultExceptionFactory(), []);
     }
 
     /**
@@ -71,7 +71,7 @@ class AbstractHydratorTest extends PHPUnit_Framework_TestCase
         ];
 
         $hydrator = $this->createHydrator("elephant");
-        $domainObject = $hydrator->hydrateForUpdate($this->createRequest($body), new ExceptionFactory(), []);
+        $domainObject = $hydrator->hydrateForUpdate($this->createRequest($body), new DefaultExceptionFactory(), []);
         $this->assertEquals([], $domainObject);
     }
 
@@ -96,7 +96,7 @@ class AbstractHydratorTest extends PHPUnit_Framework_TestCase
         ];
 
         $hydrator = $this->createHydrator("elephant", $attributeHydrator);
-        $domainObject = $hydrator->hydrateForUpdate($this->createRequest($body), new ExceptionFactory(), []);
+        $domainObject = $hydrator->hydrateForUpdate($this->createRequest($body), new DefaultExceptionFactory(), []);
         $this->assertEquals(['height' => null], $domainObject);
     }
 
@@ -121,7 +121,7 @@ class AbstractHydratorTest extends PHPUnit_Framework_TestCase
         ];
 
         $hydrator = $this->createHydrator("elephant", $attributeHydrator);
-        $domainObject = $hydrator->hydrateForUpdate($this->createRequest($body), new ExceptionFactory(), []);
+        $domainObject = $hydrator->hydrateForUpdate($this->createRequest($body), new DefaultExceptionFactory(), []);
         $this->assertEquals([], $domainObject);
     }
 
@@ -147,7 +147,7 @@ class AbstractHydratorTest extends PHPUnit_Framework_TestCase
         ];
 
         $hydrator = $this->createHydrator("elephant", $attributeHydrator);
-        $domainObject = $hydrator->hydrateForUpdate($this->createRequest($body), new ExceptionFactory(), []);
+        $domainObject = $hydrator->hydrateForUpdate($this->createRequest($body), new DefaultExceptionFactory(), []);
         $this->assertEquals(["weight" => $weight], $domainObject);
     }
 
@@ -174,7 +174,7 @@ class AbstractHydratorTest extends PHPUnit_Framework_TestCase
         ];
 
         $hydrator = $this->createHydrator("elephant", $attributeHydrator);
-        $domainObject = $hydrator->hydrateForUpdate($this->createRequest($body), new ExceptionFactory(), []);
+        $domainObject = $hydrator->hydrateForUpdate($this->createRequest($body), new DefaultExceptionFactory(), []);
         $this->assertEquals(["weight" => $weight], $domainObject);
     }
 
@@ -199,7 +199,7 @@ class AbstractHydratorTest extends PHPUnit_Framework_TestCase
         ];
 
         $hydrator = $this->createHydrator("elephant", [], $relationshipHydrator);
-        $domainObject = $hydrator->hydrateForUpdate($this->createRequest($body), new ExceptionFactory(), []);
+        $domainObject = $hydrator->hydrateForUpdate($this->createRequest($body), new DefaultExceptionFactory(), []);
         $this->assertEquals([], $domainObject);
     }
 
@@ -230,7 +230,7 @@ class AbstractHydratorTest extends PHPUnit_Framework_TestCase
         ];
 
         $hydrator = $this->createHydrator("elephant", [], $relationshipHydrator);
-        $hydrator->hydrateForUpdate($this->createRequest($body), new ExceptionFactory(), []);
+        $hydrator->hydrateForUpdate($this->createRequest($body), new DefaultExceptionFactory(), []);
     }
 
     /**
@@ -262,7 +262,7 @@ class AbstractHydratorTest extends PHPUnit_Framework_TestCase
         ];
 
         $hydrator = $this->createHydrator("elephant", [], $relationshipHydrator);
-        $hydrator->hydrateForUpdate($this->createRequest($body), new ExceptionFactory(), []);
+        $hydrator->hydrateForUpdate($this->createRequest($body), new DefaultExceptionFactory(), []);
     }
 
     /**
@@ -293,7 +293,7 @@ class AbstractHydratorTest extends PHPUnit_Framework_TestCase
         ];
 
         $hydrator = $this->createHydrator("elephant", [], $relationshipHydrator);
-        $domainObject = $hydrator->hydrateForUpdate($this->createRequest($body), new ExceptionFactory(), []);
+        $domainObject = $hydrator->hydrateForUpdate($this->createRequest($body), new DefaultExceptionFactory(), []);
         $this->assertEquals(["children" => "Dumbo"], $domainObject);
     }
 
@@ -335,7 +335,7 @@ class AbstractHydratorTest extends PHPUnit_Framework_TestCase
         ];
 
         $hydrator = $this->createHydrator("elephant", [], $relationshipHydrator);
-        $domainObject = $hydrator->hydrateForUpdate($this->createRequest($body), new ExceptionFactory(), []);
+        $domainObject = $hydrator->hydrateForUpdate($this->createRequest($body), new DefaultExceptionFactory(), []);
         $this->assertEquals(["owner" => "1", "children" => ["2"]], $domainObject);
     }
 
@@ -347,7 +347,7 @@ class AbstractHydratorTest extends PHPUnit_Framework_TestCase
             ->withBody(new Stream("php://memory", "rw"));
         $psrRequest->getBody()->write(json_encode($body));
 
-        $request = new Request($psrRequest, new ExceptionFactory());
+        $request = new Request($psrRequest, new DefaultExceptionFactory());
 
         return $request;
     }