+++ tests/JsonApi/Response/ResponderTest.php
@@ -3,11 +3,12 @@ namespace WoohooLabsTest\Yin\JsonApi\Response;
 
 use PHPUnit_Framework_TestCase;
 use WoohooLabs\Yin\JsonApi\Document\ErrorDocument;
-use WoohooLabs\Yin\JsonApi\Exception\ExceptionFactory;
+use WoohooLabs\Yin\JsonApi\Exception\DefaultExceptionFactory;
 use WoohooLabs\Yin\JsonApi\Request\Request;
 use WoohooLabs\Yin\JsonApi\Response\Responder;
 use WoohooLabs\Yin\JsonApi\Schema\Link;
 use WoohooLabs\Yin\JsonApi\Schema\Links;
+use WoohooLabs\Yin\JsonApi\Serializer\DefaultSerializer;
 use WoohooLabsTest\Yin\JsonApi\Utils\StubSuccessfulDocument;
 use Zend\Diactoros\Response;
 use Zend\Diactoros\ServerRequest;
@@ -114,9 +115,10 @@ class ResponderTest extends PHPUnit_Framework_TestCase
     private function createResponder()
     {
         return new Responder(
-            new Request(new ServerRequest(), new ExceptionFactory()),
+            new Request(new ServerRequest(), new DefaultExceptionFactory()),
             new Response(),
-            new ExceptionFactory()
+            new DefaultExceptionFactory(),
+            new DefaultSerializer()
         );
     }
 }